+++ LinkEdit.mm
@@ -30,8 +30,8 @@ using namespace std;
 //-----------------------------------------------------------------------------
 - (MVNode *) createRelocNode:(MVNode *)parent
                      caption:(NSString *)caption
-                    location:(uint32_t)location
-                      length:(uint32_t)length
+                    location:(uint64_t)location
+                      length:(uint64_t)length
                  baseAddress:(uint32_t)baseAddress // start of the section containing the relocation (image: start of the first segment)
 {
   MVNodeSaver nodeSaver;
@@ -82,7 +82,7 @@ using namespace std;
     
     if (relocation_info)
     {
-      uint32_t relocLocation = [self RVAToFileOffset:baseAddress + relocation_info->r_address];
+      uint64_t relocLocation = [self RVAToFileOffset:baseAddress + relocation_info->r_address];
       NSRange rangeReloc = NSMakeRange(relocLocation,0);
       uint32_t relocValue = [dataController read_uint32:rangeReloc];
       uint32_t relocLength = (1 << relocation_info->r_length);
@@ -206,7 +206,7 @@ using namespace std;
                )
       {
         //read original content at relocation
-        uint32_t relocLocation = [self RVAToFileOffset:baseAddress + prev_scattered_relocation_info->r_address];
+        uint64_t relocLocation = [self RVAToFileOffset:baseAddress + prev_scattered_relocation_info->r_address];
         uint32_t relocLength = (1 << prev_scattered_relocation_info->r_length);
         NSAssert1(relocLength == sizeof(uint32_t), @"unsupported reloc length (%u)", relocLength);
         NSRange rangeReloc = NSMakeRange(relocLocation,0);
@@ -254,11 +254,11 @@ using namespace std;
                               r_type == ARM_RELOC_PAIR ? @"ARM_RELOC_PAIR" :
                               r_type == ARM_RELOC_SECTDIFF ? @"ARM_RELOC_SECTDIFF" :
                               r_type == ARM_RELOC_LOCAL_SECTDIFF ? @"ARM_RELOC_LOCAL_SECTDIFF" :
-                              r_type == ARM_RELOC_PB_LA_PTR ? @"ARM_RELOC_PB_LA_PTR" : 
+                              r_type == ARM_RELOC_PB_LA_PTR ? @"ARM_RELOC_PB_LA_PTR" :
                               r_type == ARM_RELOC_BR24 ? @"ARM_RELOC_BR24" :
-                              r_type == ARM_THUMB_RELOC_BR22 ? @"ARM_THUMB_RELOC_BR22" : 
+                              r_type == ARM_THUMB_RELOC_BR22 ? @"ARM_THUMB_RELOC_BR22" :
                               r_type == ARM_THUMB_32BIT_BRANCH ? @"ARM_THUMB_32BIT_BRANCH" :
-                              r_type == ARM_RELOC_HALF ? @"ARM_RELOC_HALF" : 
+                              r_type == ARM_RELOC_HALF ? @"ARM_RELOC_HALF" :
                               r_type == ARM_RELOC_HALF_SECTDIFF ? @"ARM_RELOC_HALF_SECTDIFF" : @"?????"];
     }
     
@@ -297,8 +297,8 @@ using namespace std;
 //-----------------------------------------------------------------------------
 - (MVNode *) createReloc64Node:(MVNode *)parent
                        caption:(NSString *)caption
-                      location:(uint32_t)location
-                        length:(uint32_t)length
+                      location:(uint64_t)location
+                        length:(uint64_t)length
                    baseAddress:(uint64_t)baseAddress // start of the section containing the relocation (image: start of the first segment)
 {
   MVNodeSaver nodeSaver;
@@ -342,7 +342,7 @@ using namespace std;
     //========================================================================
     if (relocation_info->r_extern)
     {
-      uint32_t relocLocation = [self RVAToFileOffset:baseAddress + relocation_info->r_address];
+      uint64_t relocLocation = [self RVAToFileOffset:baseAddress + relocation_info->r_address];
       NSRange rangeReloc = NSMakeRange(relocLocation,0);
 
       // target symbol
@@ -409,7 +409,7 @@ using namespace std;
                                     ? [NSString stringWithFormat:@"-0x%X",-relocAddend] 
                                     : [NSString stringWithFormat:@"0x%X",relocAddend]];
           }
-          uint32_t relocValue = nlist_64->n_value - prev_nlist_64->n_value + relocAddend;
+          uint64_t relocValue = nlist_64->n_value - prev_nlist_64->n_value + relocAddend;
           
           // update real data
           [self addRelocAtFileOffset:relocLocation withLength:relocLength andValue:relocValue];
@@ -442,7 +442,8 @@ using namespace std;
         {
           // 32bit signed PC Rel
           NSParameterAssert(relocation_info->r_pcrel == true);
-          uint32_t relocValue = nlist_64->n_value - relocation_info->r_address - baseAddress - relocLength;
+          // XXX: is this cast correct?
+          uint32_t relocValue = (uint32_t)(nlist_64->n_value - relocation_info->r_address - baseAddress - relocLength);
           uint32_t relocAddend = [dataController read_uint32:rangeReloc];
 
           if (mach_header_64->cputype == CPU_TYPE_X86_64)
@@ -512,7 +513,8 @@ using namespace std;
                                     ? [NSString stringWithFormat:@"-0x%X",-relocAddend] 
                                     : [NSString stringWithFormat:@"0x%X",relocAddend]];
           }
-          uint32_t relocValue = *symbols_64.begin() - nlist_64 - 1;
+          // XXX: is this cast correct?
+          uint32_t relocValue = (uint32_t)(*symbols_64.begin() - nlist_64 - 1);
           relocValue -= relocation_info->r_address + baseAddress + relocLength; // it is PC relative
           
           // update real data
@@ -571,7 +573,7 @@ using namespace std;
                                :@"Section"
                                :[NSString stringWithFormat:@"%u %@", relocation_info->r_symbolnum, sectionName]];
         
-        uint32_t relocLocation = [self RVAToFileOffset:baseAddress + relocation_info->r_address];
+        uint64_t relocLocation = [self RVAToFileOffset:baseAddress + relocation_info->r_address];
         NSRange rangeReloc = NSMakeRange(relocLocation,0);
         uint64_t relocValue = 0;
         
@@ -692,8 +694,8 @@ using namespace std;
 //-----------------------------------------------------------------------------
 - (MVNode *) createSymbolsNode:parent 
                        caption:(NSString *)caption
-                      location:(uint32_t)location
-                        length:(uint32_t)length
+                      location:(uint64_t)location
+                        length:(uint64_t)length
 {
   MVNodeSaver nodeSaver;
   MVNode * node = [parent insertChildWithDetails:caption location:location length:length saver:nodeSaver]; 
@@ -846,7 +848,7 @@ using namespace std;
                              :[NSString stringWithFormat:@"%u", nlist->n_value]];
       
       // fill in lookup table with undefined sybols (key equals (-1) * index)
-      uint32_t key = *symbols.begin() - nlist - 1;
+      uint32_t key = (uint32_t)(*symbols.begin() - nlist - 1);
       [symbolNames setObject:symbolName
                       forKey:[NSNumber numberWithUnsignedLong:key]];
     }
@@ -862,8 +864,8 @@ using namespace std;
 //-----------------------------------------------------------------------------
 - (MVNode *) createSymbols64Node:parent 
                          caption:(NSString *)caption
-                        location:(uint32_t)location
-                          length:(uint32_t)length
+                        location:(uint64_t)location
+                          length:(uint64_t)length
 {
   MVNodeSaver nodeSaver;
   MVNode * node = [parent insertChildWithDetails:caption location:location length:length saver:nodeSaver]; 
@@ -1026,8 +1028,8 @@ using namespace std;
 //-----------------------------------------------------------------------------
 - (MVNode *) createReferencesNode:parent 
                           caption:(NSString *)caption
-                         location:(uint32_t)location
-                           length:(uint32_t)length
+                         location:(uint64_t)location
+                           length:(uint64_t)length
 {
   MVNodeSaver nodeSaver;
   MVNode * node = [parent insertChildWithDetails:caption location:location length:length saver:nodeSaver]; 
@@ -1080,8 +1082,8 @@ using namespace std;
 //-----------------------------------------------------------------------------
 - (MVNode *) createISymbolsNode:parent
                         caption:(NSString *)caption
-                       location:(uint32_t)location
-                         length:(uint32_t)length
+                       location:(uint64_t)location
+                         length:(uint64_t)length
 {
   MVNodeSaver nodeSaver;
   MVNode * node = [parent insertChildWithDetails:caption location:location length:length saver:nodeSaver]; 
@@ -1091,7 +1093,7 @@ using namespace std;
   
   for (uint32_t nindsym = 0; nindsym < length / sizeof(uint32_t); ++nindsym)
   {
-    uint32_t nsect = sections.size();
+    uint32_t nsect = (uint32_t)sections.size();
     while (--nsect > 0)
     {
       struct section const * section = [self getSectionByIndex:nsect];
@@ -1208,8 +1210,8 @@ using namespace std;
 //-----------------------------------------------------------------------------
 - (MVNode *) createISymbols64Node:parent
                           caption:(NSString *)caption
-                         location:(uint32_t)location
-                           length:(uint32_t)length
+                         location:(uint64_t)location
+                           length:(uint64_t)length
 {
   MVNodeSaver nodeSaver;
   MVNode * node = [parent insertChildWithDetails:caption location:location length:length saver:nodeSaver]; 
@@ -1219,7 +1221,7 @@ using namespace std;
   
   for (uint32_t nindsym = 0; nindsym < length / sizeof(uint32_t); ++nindsym)
   {
-    uint32_t nsect = sections_64.size();
+    uint32_t nsect = (uint32_t)sections_64.size();
     while (--nsect > 0)
     {
       struct section_64 const * section_64 = [self getSection64ByIndex:nsect];
@@ -1239,7 +1241,7 @@ using namespace std;
 
       // calculate stub or pointer length
       uint32_t length = (section_64->reserved2 > 0 ? section_64->reserved2 : sizeof(uint64_t));
-        
+
       // calculate indirect value location
       uint64_t indirectAddress = section_64->addr + (nindsym - section_64->reserved1) * length;
         
@@ -1336,8 +1338,8 @@ using namespace std;
 //-----------------------------------------------------------------------------
 - (MVNode *) createTOCNode:parent
                    caption:(NSString *)caption
-                  location:(uint32_t)location
-                    length:(uint32_t)length
+                  location:(uint64_t)location
+                    length:(uint64_t)length
 {
   MVNodeSaver nodeSaver;
   MVNode * node = [parent insertChildWithDetails:caption location:location length:length saver:nodeSaver]; 
@@ -1387,8 +1389,8 @@ using namespace std;
 //-----------------------------------------------------------------------------
 - (MVNode *) createTOC64Node:parent
                      caption:(NSString *)caption
-                    location:(uint32_t)location
-                      length:(uint32_t)length
+                    location:(uint64_t)location
+                      length:(uint64_t)length
 {
   MVNodeSaver nodeSaver;
   MVNode * node = [parent insertChildWithDetails:caption location:location length:length saver:nodeSaver]; 
@@ -1438,8 +1440,8 @@ using namespace std;
 //-----------------------------------------------------------------------------
 - (MVNode *) createModulesNode:parent
                        caption:(NSString *)caption
-                      location:(uint32_t)location
-                        length:(uint32_t)length
+                      location:(uint64_t)location
+                        length:(uint64_t)length
 {
   MVNodeSaver nodeSaver;
   MVNode * node = [parent insertChildWithDetails:caption location:location length:length saver:nodeSaver]; 
@@ -1550,8 +1552,8 @@ using namespace std;
 //-----------------------------------------------------------------------------
 - (MVNode *) createModules64Node:parent
                          caption:(NSString *)caption
-                        location:(uint32_t)location
-                          length:(uint32_t)length
+                        location:(uint64_t)location
+                          length:(uint64_t)length
 {
   MVNodeSaver nodeSaver;
   MVNode * node = [parent insertChildWithDetails:caption location:location length:length saver:nodeSaver]; 
@@ -1662,8 +1664,8 @@ using namespace std;
 //-----------------------------------------------------------------------------
 - (MVNode *) createTwoLevelHintsNode:parent 
                              caption:(NSString *)caption
-                            location:(uint32_t)location
-                              length:(uint32_t)length
+                            location:(uint64_t)location
+                              length:(uint64_t)length
                                index:(uint32_t)index
 {
   MVNodeSaver nodeSaver;
@@ -1710,8 +1712,8 @@ using namespace std;
 //-----------------------------------------------------------------------------
 - (MVNode *) createSplitSegmentNode:parent
                             caption:(NSString *)caption
-                           location:(uint32_t)location
-                             length:(uint32_t)length
+                           location:(uint64_t)location
+                             length:(uint64_t)length
                         baseAddress:(uint64_t)baseAddress
 {
   MVNodeSaver nodeSaver;
@@ -1785,8 +1787,8 @@ using namespace std;
 //-----------------------------------------------------------------------------
 - (MVNode *) createFunctionStartsNode:parent
                               caption:(NSString *)caption
-                             location:(uint32_t)location
-                               length:(uint32_t)length
+                             location:(uint64_t)location
+                               length:(uint64_t)length
                           baseAddress:(uint64_t)baseAddress
 {
   MVNodeSaver nodeSaver;
@@ -1819,8 +1821,8 @@ using namespace std;
 //-----------------------------------------------------------------------------
 - (MVNode *) createDataInCodeEntriesNode:parent
                                  caption:(NSString *)caption
-                                location:(uint32_t)location
-                                  length:(uint32_t)length
+                                location:(uint64_t)location
+                                  length:(uint64_t)length
 {
   MVNodeSaver nodeSaver;
   MVNode * node = [parent insertChildWithDetails:caption location:location length:length saver:nodeSaver];
